requirements.txt
langchain
langgraph
openai
sqlalchemy
psycopg2-binary
pandas
python-dotenv
streamlit
langchain-openai
langchain-community